requirements.txt
openai-whisper
torch
torchaudio
TTS
ollama
fastapi
uvicorn[standard]
python-multipart
